+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add support for socket files to the `Proxy` site type
 * Domain names can now contain named capture groups, which can be used in the socket file
 * The `HOME` environment variable will be unset if a node site runs as a different UID
+* Add `fallback_file` option to static site settings
 
 ## 0.4.0 (2020-07-21)